chore(transaction-pay-controller): relocate chomp utility to utils/
